refactor: consolidate utility functions imports
Update test cases to use `filePathUtilities` object for consistency
and clarity. Change import statements from `windowUtilities` to
`promptUtilities` in `configurationHandlers` tests to reflect the
correct utility function origins.

- Consolidate multiple imports into a single import for
  `filePathUtilities`.
- Update stubs and assertions in `filePathUtilities` tests.
- Adjust mock key value in `nl/common.json` test for consistency.
- Introduce `setup` and `teardown` in `i18nextJsonToPoConversionModule`
  tests for better configuration management.
