+++ src/lib/modules/i18nextJsonToPoConversion/i18nextJsonToPoConversionModule.ts
@@ -1,7 +1,9 @@
 import * as Sentry from '@sentry/node';
 import { i18next2po } from 'gettext-converter';
 
+import I18nextJsonToPoConversionModuleConfiguration from '../../entities/configuration/modules/I18nextJsonToPoConversionModule/i18nextJsonToPoConversionModuleConfiguration';
 import FileWriter from '../../services/fileIo/fileWriter';
+import ConfigurationStoreManager from '../../stores/configuration/configurationStoreManager';
 import { BaseActionModule } from '../baseActionModule';
 import I18nextJsonToPoConversionModuleContext from './i18nextJsonToPoConversionModuleContext';
 
@@ -17,30 +19,36 @@ export default class I18nextJsonToPoConversionModule extends BaseActionModule {
   protected async doExecuteAsync(
     context: I18nextJsonToPoConversionModuleContext
   ): Promise<void> {
-    await Sentry.startSpan(
-      {
-        op: 'json.convertJsonToPo',
-        name: 'JSON to PO Conversion Module',
-      },
-      async () => {
-        console.log(
-          `Converting json to po using : ${context.inputPath.fsPath}`
-        );
-        if (context.jsonContent) {
-          try {
-            const res = i18next2po(context.locale, context.jsonContent, {
-              compatibilityJSON: 'v3',
-            });
-            await FileWriter.writeToFileAsync(context.outputPath, res);
-          } catch (error) {
-            if (error instanceof SyntaxError) {
-              Sentry.captureException(error);
-            } else {
-              Sentry.captureException(error, { extra: { context } });
+    if (
+      ConfigurationStoreManager.getInstance().getConfig<I18nextJsonToPoConversionModuleConfiguration>(
+        'i18nextJsonToPoConversionModule'
+      ).enabled
+    ) {
+      await Sentry.startSpan(
+        {
+          op: 'json.convertJsonToPo',
+          name: 'JSON to PO Conversion Module',
+        },
+        async () => {
+          console.log(
+            `Converting json to po using : ${context.inputPath.fsPath}`
+          );
+          if (context.jsonContent) {
+            try {
+              const res = i18next2po(context.locale, context.jsonContent, {
+                compatibilityJSON: 'v3',
+              });
+              await FileWriter.writeToFileAsync(context.outputPath, res);
+            } catch (error) {
+              if (error instanceof SyntaxError) {
+                Sentry.captureException(error);
+              } else {
+                Sentry.captureException(error, { extra: { context } });
+              }
             }
           }
         }
-      }
-    );
+      );
+    }
   }
 }
