+++ src/lib/modules/poToI18nextJsonConversion/poToI18nextJsonConversionModule.ts
@@ -2,7 +2,9 @@ import * as Sentry from '@sentry/node';
 import { po2i18next } from 'gettext-converter';
 import stringify from 'json-stable-stringify';
 
+import I18nextJsonToPoConversionModuleConfiguration from '../../entities/configuration/modules/I18nextJsonToPoConversionModule/i18nextJsonToPoConversionModuleConfiguration';
 import FileWriter from '../../services/fileIo/fileWriter';
+import ConfigurationStoreManager from '../../stores/configuration/configurationStoreManager';
 import { BaseActionModule } from '../baseActionModule';
 import PoToI18nextJsonConversionModuleContext from './poToI18nextJsonConversionModuleContext';
 
@@ -18,29 +20,35 @@ export default class PoToI18nextJsonConversionModule extends BaseActionModule {
   protected async doExecuteAsync(
     context: PoToI18nextJsonConversionModuleContext
   ): Promise<void> {
-    await Sentry.startSpan(
-      {
-        op: 'po.convertPoToJson',
-        name: 'PO to JSON Conversion Module',
-      },
-      async () => {
-        console.log(
-          `Converting po to json using : ${context.inputPath.fsPath}`
-        );
-        if (context.poContent) {
-          const res = po2i18next(context.poContent, {
-            compatibilityJSON: 'v3',
-          });
+    if (
+      ConfigurationStoreManager.getInstance().getConfig<I18nextJsonToPoConversionModuleConfiguration>(
+        'i18nextJsonToPoConversionModule'
+      ).enabled
+    ) {
+      await Sentry.startSpan(
+        {
+          op: 'po.convertPoToJson',
+          name: 'PO to JSON Conversion Module',
+        },
+        async () => {
+          console.log(
+            `Converting po to json using : ${context.inputPath.fsPath}`
+          );
+          if (context.poContent) {
+            const res = po2i18next(context.poContent, {
+              compatibilityJSON: 'v3',
+            });
 
-          let jsonResult = stringify(res, {
-            space: 4,
-            cycles: false,
-          });
-          jsonResult = jsonResult + '\n';
+            let jsonResult = stringify(res, {
+              space: 4,
+              cycles: false,
+            });
+            jsonResult = jsonResult + '\n';
 
-          await FileWriter.writeToFileAsync(context.outputPath, jsonResult);
+            await FileWriter.writeToFileAsync(context.outputPath, jsonResult);
+          }
         }
-      }
-    );
+      );
+    }
   }
 }
