+++ src/lib/utilities/promptUtilities.test.ts
@@ -9,9 +9,20 @@ import {
   selectFrameworkAsync,
   selectProjectTypeAsync,
   showConfigurationToUserAsync,
+  showOpenDialog,
 } from './promptUtilities';
 
 suite('promptUtilities', () => {
+  let sandbox: sinon.SinonSandbox;
+
+  setup(() => {
+    sandbox = sinon.createSandbox();
+  });
+
+  teardown(() => {
+    sandbox.restore();
+  });
+
   suite('selectProjectTypeAsync', () => {
     test('should return selected project type', async () => {
       const projectTypes = Object.values(ProjectType);
@@ -68,5 +79,45 @@ suite('promptUtilities', () => {
 
       showInformationMessageStub.restore();
     });
+  });
+
+  test('showOpenDialog should return undefined if folderSelectionPrompt is falsy', async () => {
+    sandbox.stub(vscode.window, 'showQuickPick').resolves(undefined);
+
+    const result = await showOpenDialog('Select Folder');
+
+    assert.strictEqual(result, undefined);
+  });
+
+  test('showOpenDialog should return undefined if folderUri is undefined', async () => {
+    sandbox.stub(vscode.window, 'showQuickPick').resolves({} as any);
+    sandbox.stub(vscode.window, 'showOpenDialog').resolves(undefined);
+
+    const result = await showOpenDialog('Select Folder');
+
+    assert.strictEqual(result, undefined);
+  });
+
+  test('showOpenDialog should return and array with length one if only a single folder was picked', async () => {
+    const folderUri = [vscode.Uri.file('/path/to/folder')];
+    sandbox.stub(vscode.window, 'showQuickPick').resolves({} as any);
+    sandbox.stub(vscode.window, 'showOpenDialog').resolves(folderUri);
+
+    const result = await showOpenDialog('Select Folder');
+
+    assert.deepEqual(result, ['/path/to/folder']);
+  });
+
+  test('showOpenDialog should return an array of folderUri.fsPath if folderUri has length greater than 1', async () => {
+    const folderUri = [
+      vscode.Uri.file('/path/to/folder1'),
+      vscode.Uri.file('/path/to/folder2'),
+    ];
+    sandbox.stub(vscode.window, 'showQuickPick').resolves({} as any);
+    sandbox.stub(vscode.window, 'showOpenDialog').resolves(folderUri);
+
+    const result = await showOpenDialog('Select Folder', true);
+
+    assert.deepStrictEqual(result, ['/path/to/folder1', '/path/to/folder2']);
   });
 });
