+++ src/test/mock-workspace/locales/nl/common.po
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2024-07-21T11:11:09.289Z\n"
-"PO-Revision-Date: 2024-07-21T11:11:09.289Z\n"
+"POT-Creation-Date: 2024-07-22T17:49:47.266Z\n"
+"PO-Revision-Date: 2024-07-22T17:49:47.266Z\n"
 "Language: en-US\n"
 
 msgid "key"
